Refactor & code cleanup
